+++ generators/page/templates/gulp/tasks/aws.js
@@ -10,7 +10,7 @@ const cloudfront = require('gulp-cloudfront-invalidate-aws-publish');
 const confirm = require('gulp-confirm');
 const deline = require('deline');
 const gulp = require('gulp');
-const gutil = require('gulp-util');
+const log = require('fancy-log');
 const rename = require('gulp-rename');
 
 const awsJson = require('./../../aws.json');
@@ -69,11 +69,6 @@ module.exports = () =>
         .pipe(cloudfront(cfSettings))
         .pipe(publisher.cache())
         .pipe(awspublish.reporter())
-        .on(
-          'end',
-          gutil.log.bind(
-            gutil,
-            // eslint-disable-next-line comma-dangle
-            `Published at 'http://interactives.dallasnews.com/${awsDirectory}'.`
-          )  // eslint-disable-line comma-dangle
-        );
+        .on('end', () => {
+          log(`Published at 'http://interactives.dallasnews.com/${awsDirectory}'.`);
+        });